requirements.txt
alembic==1.4.1
boto3==1.17.110
botocore==1.20.110
certifi==2021.5.30
charset-normalizer==2.0.4
click==8.0.1
cloudpickle==1.6.0
databricks-cli==0.15.0
dataclasses==0.6
docker==5.0.0
entrypoints==0.3
Flask==2.0.1
future==0.18.2
gitdb==4.0.7
GitPython==3.1.18
gunicorn==20.1.0
idna==3.2
importlib-metadata==4.6.4
itsdangerous==2.0.1
Jinja2==3.0.1
jmespath==0.10.0
Mako==1.1.5
MarkupSafe==2.0.1
mlflow==1.20.1
numpy==1.21.2
packaging==21.0
pandas==1.1.5
picamera==1.13
Pillow==8.3.1
prometheus-client==0.11.0
prometheus-flask-exporter==0.18.2
protobuf==3.17.3
pyparsing==2.4.7
python-dateutil==2.8.1
python-editor==1.0.4
pytz==2021.1
PyYAML==5.4.1
querystring-parser==1.2.4
requests==2.26.0
s3transfer==0.4.2
six==1.16.0
smmap==4.0.0
SQLAlchemy==1.4.23
sqlparse==0.4.1
tabulate==0.8.9
typing-extensions==3.10.0.0
urllib3==1.26.6
websocket-client==1.2.1
Werkzeug==2.0.1
zipp==3.5.0

